Cater for log file not beign created

+++ automation/remote/logWatch.sh
@@ -48,23 +48,27 @@ lastLineNumber=0
 exitCode=4366
 while [ $retryCount -le $retryMax ] && [ $exitCode -ne 0 ]; do
 	sleep $wait
-	output=$(cat $logFile)
-	if [ -z "$output" ]; then
-		echo "[$scriptName]   no output ..."
-    else
-		lineCount=1
-		while read -r line; do
-	    	if [ $lineCount -gt $lastLineNumber ]; then
-				echo "> $line"
-				lastLineNumber=$lineCount
-			fi	
-			let "lineCount=lineCount+1"
-		done < <(echo "$output")
-	
-		found=$(echo $output | grep "$stringMatch")
-	    if [ ! -z "$found" ]; then
-			echo "[$scriptName] stringMatch ($stringMatch) found."
-		    exitCode=0
+	if [ ! -f $logFile ]; then
+		echo "[$scriptName]   logfile $logFile does not exist ..."
+	else
+		output=$(cat $logFile)
+		if [ -z "$output" ]; then
+			echo "[$scriptName]   no output ..."
+	    else
+			lineCount=1
+			while read -r line; do
+		    	if [ $lineCount -gt $lastLineNumber ]; then
+					echo "> $line"
+					lastLineNumber=$lineCount
+				fi	
+				let "lineCount=lineCount+1"
+			done < <(echo "$output")
+		
+			found=$(echo $output | grep "$stringMatch")
+		    if [ ! -z "$found" ]; then
+				echo "[$scriptName] stringMatch ($stringMatch) found."
+			    exitCode=0
+			fi
 		fi
 	fi
